player health is float- fixed method

+++ core/src/com/hackslash/game/model/Player.java
@@ -83,7 +83,7 @@ public class Player extends GameObject {
         player_position.y = dy;
     }
 
-    public int getPlayerHealth(){
+    public float getPlayerHealth(){
         return playerHealth;
